Mark the symbol '__isr_vectors' undefined.
This is mandatory if the system folder is linked as a library.
Since the __isr_vector symbol is never referenced and the linker only
includes referenced symbols from libraries.
If omitted, the symbol never loaded, which leaves the '.isr_vector' section empty.

The downsize of this: the fixed '__isr_vector' symbol name.
Furthermore: all this also holds for the '.cfmconfig' section.


Signed-off-by: Ben Wijen <ben@wijen.net>